fix: separate integration and unit tests in config for market-data-simulator-go
